requirements.txt
streamlit
streamlit-extras
pandas
plotly
PyYAML
requests
beautifulsoup4
xlsxwriter
kaleido
